refactored utils.hpp to locks.hpp
